refactor(computer-tool): split by responsibility
